Refactoring 'utils.average' into an own module 'average'
