+++ scripts/pyFAI-average
@@ -30,7 +30,7 @@ __author__ = "Jerome Kieffer, Picca Frédéric-Emmanuel"
 __contact__ = "Jerome.Kieffer@ESRF.eu"
 __license__ = "GPLv3+"
 __copyright__ = "European Synchrotron Radiation Facility, Grenoble, France"
-__date__ = "17/06/2016"
+__date__ = "27/07/2016"
 __status__ = "production"
 
 import os
@@ -124,7 +124,7 @@ def main():
     else:
         quantiles = None
     if images:
-        dataout = pyFAI.utils.averageImages(images, filter_=options.method, cutoff=options.cutoff,
+        dataout = pyFAI.average.average_images(images, filter_=options.method, cutoff=options.cutoff,
                                             threshold=0, fformat=options.format, output=output,
                                             flats=flats, darks=darks, quantiles=quantiles)
         return dataout
